finish history log be without searchbar

+++ app/src/main/java/com/calorify/app/data/local/PieChartData.kt
@@ -2,5 +2,5 @@ package com.calorify.app.data.local
 
 data class PieChartData(
     var status: String?,
-    var value: Float?
+    var value: Int?
 )
