+++ app/src/main/java/com/calorify/app/ui/screen/profile/SelfAssessmentResultScreen.kt
@@ -6,11 +6,9 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import com.calorify.app.databinding.AssessmentResultScreenBinding
-import com.calorify.app.ui.theme.CalorifyTheme
 
 @Composable
 fun SelfAssessmentResultScreen(
@@ -18,7 +16,7 @@ fun SelfAssessmentResultScreen(
     height: String,
     indexBmi: Float,
     weightGoal: String,
-    calorie: Float,
+    calorie: Int,
     onDoAssessmentClick: () -> Unit,
 ) {
     AndroidView(
